+++ xxcmd/__init__.py
@@ -39,6 +39,10 @@ def main():
         help="Don't echo the command to the terminal prior to execution.")
 
     parser.add_argument(
+        '-p', '--label-padding', action='store', metavar='PADDING', type=int,
+        help="Add extra padding between labels and commands.")
+
+    parser.add_argument(
         '-t', '--no-labels', action='store_true',
         help="Don't display command labels.")
 
@@ -69,6 +73,8 @@ def main():
     manager.config.show_labels = not args.no_labels
     manager.config.echo_commands = not args.no_echo
     manager.config.draw_window_border = not args.no_border
+    if args.label_padding:
+        manager.config.label_padding = args.label_padding
     # Load db
     manager.load_database()