+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added support for scrolling through long command lists.
 - Added visual border around the window.
 - Added option to not show the window border.
+- Added option to increase padding between labels and commands.
 
 ## [0.4.0] - 2021-02-08
 - Added option to not echo commands when executed (-n, --no-echo)